Migrate the build configuration from Groovy to KTS
Migrate app level and project level build
configuration from Groovy toKTS(kotlin script).
KTS and kotlin DSL are the same thing.
The migration reason is because kotlin DSL allow
for control click and navigates to object definitions.

Also move const dependencies and versions
to be inside the buildSrc. This will allow
us to share the usage of all const dependencies
and versions in the future. Also make updating
the version number much easier due to the fact
that we only need to change version in one spot.

Follow this link for help: https://proandroiddev.com/migrate-from-groovy-to-kotlin-dsl-951266f3c072