Create repository and repositoryImpl
- having a repository layer is not
necessary but it helps keep a clean
API architecture, since it helps
isolate the data access for the
rest of the app

- the reason we have repository
interface is due to the fact that
we may potentially want different class
implementation of the same interface,
but when it comes to koin dependency
injection, we can define the the same
interface type with different definitions.
And koin is smart to know which specific
implementation to inject.