Update the data class
- add more details to some
of the data class variables,
so it is more in the desired final format,
this makes the insertion of data
to the UI layer cleaner. Such as we are not concatenating
all the hard coded strings in the UI layer. UI is
just suppose to show the view only
and not perform any more logic than that.